+++ NEWS.md
@@ -1,7 +1,9 @@
 # mlr3misc 0.1.1
 
+* New function `encapsulate()` to call functions while recording a log.
+* `invoke()`: New arguments `.opts` and `.seed` to temporarily set options or
+  random seeds, respectively.
 * Fixed warnings about partial argument matching.
-* `invoke()`: New argument `.opts` to temporarily set options.
 
 # mlr3misc 0.1.0